+++ src/projectapis/conceptual/FactorialAPIImplementation.java
@@ -1,7 +1,6 @@
 package projectapis.conceptual;
 
 import project.annotations.ConceptualAPI;
-import projectapis.FactorialAPI;
 
 @ConceptualAPI
 public class FactorialAPIImplementation implements FactorialAPI {
