+++ src/projectapis/conceptual/FactorialAPIPrototype.java
@@ -1,7 +1,6 @@
 package projectapis.conceptual;
 
 import project.annotations.ConceptualAPIPrototype;
-import projectapis.FactorialAPI;
 
 public class FactorialAPIPrototype {
