+++ src/projectapis/network/UserAPIPrototype.java
@@ -1,7 +1,6 @@
 package projectapis.network;
 
 import project.annotations.NetworkAPIPrototype;
-import projectapis.UserAPI;
 
 public class UserAPIPrototype {
