+++ test/smoketests/DataStorageAPITest.java
@@ -3,8 +3,8 @@ package smoketests;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 
+import projectapis.process.DataStorageAPI;
 import projectapis.process.DataStorageAPIImplementation;
-import projectapis.DataStorageAPI;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
